requirements.txt
Django==4.1
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
asgiref==3.6.0
sqlparse==0.4.4
